+++ treebeard/__init__.py
@@ -18,4 +18,4 @@ Release logic:
 14. git push
 """
 
-__version__ = "5.1.0"
+__version__ = "5.2.0"